consolidate content page into single source of truth
- merge admin and employee content views into one ContentPage component
- delete apps/web/src/pages/admin/content/page.tsx and role-content.tsx
- update App.tsx to import ContentPage directly
- fix admin view missing tags and checkout badge
- sync apps/admin content page to match card UI
