+++ apps/web/src/pages/content/page.tsx
@@ -3,6 +3,12 @@ import { useState } from "react";
 import { Link } from "react-router";
 import { trpc } from "@/lib/trpc.ts";
 
+const ROLE_TABS = [
+  { key: "all", label: "All Users" },
+  { key: "underwriter", label: "Underwriter" },
+  { key: "business-analyst", label: "Business Analyst" },
+];
+
 function getStatusBadge(status: string | null) {
   switch (status) {
     case "Finalized":
@@ -18,10 +24,22 @@ function getStatusBadge(status: string | null) {
   }
 }
 
-function ContentListPage() {
+function matchesOwnerRole(owner: { role: string } | null | undefined, role: string): boolean {
+  if (role === "all") return true;
+  const r = owner?.role ?? "";
+  if (role === "underwriter") return r === "underwriter";
+  if (role === "business-analyst") return r === "business-analyst";
+  return false;
+}
+
+function ContentPage() {
+  const { data: access } = trpc.user.myAccess.useQuery();
+  const isAdmin = access?.role === "admin";
+
   const [search, setSearch] = useState("");
-  const [status, setStatus] = useState("");
-  const [type, setType] = useState("");
+  const [statusFilter, setStatusFilter] = useState("");
+  const [typeFilter, setTypeFilter] = useState("");
+  const [roleFilter, setRoleFilter] = useState("all");
 
   const utils = trpc.useUtils();
   const toggleFavorite = trpc.content.update.useMutation({
@@ -30,53 +48,82 @@ function ContentListPage() {
 
   const contents = trpc.content.list.useQuery({
     search,
-    document_status: status || undefined,
-    content_type: (type as "Reference" | "Workflow") || undefined,
+    document_status: statusFilter || undefined,
+    content_type: (typeFilter as "Reference" | "Workflow") || undefined,
   });
+
+  const allItems = contents.data ?? [];
+  const filtered = isAdmin
+    ? allItems.filter((item) => matchesOwnerRole(item.owner, roleFilter))
+    : allItems;
 
   return (
     <div id="content-library" className="scroll-mt-4 border-t border-border/60 py-6 sm:py-8">
       <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
-        <div className="mb-6 flex flex-col gap-3 sm:flex-row sm:items-end">
-          <div className="relative flex-1">
-            <Search className="absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2 text-muted-foreground" />
-            <input
-              type="text"
-              placeholder="Search by filename or URL..."
-              value={search}
-              onChange={(e) => setSearch(e.target.value)}
-              className="w-full rounded border border-border bg-background py-2 pl-10 pr-4 text-sm focus:outline-none focus:ring-2 focus:ring-hanover-green"
-            />
+        <div className="mb-6 flex flex-col gap-4">
+          {isAdmin && (
+            <div className="flex flex-wrap items-center gap-2">
+              {ROLE_TABS.map((tab) => (
+                <button
+                  key={tab.key}
+                  type="button"
+                  onClick={() => setRoleFilter(tab.key)}
+                  className={`rounded px-4 py-1.5 text-sm font-medium transition-colors ${
+                    roleFilter === tab.key
+                      ? "bg-hanover-deepblue text-white"
+                      : "border border-border bg-card text-muted-foreground hover:text-foreground"
+                  }`}
+                >
+                  {tab.label}
+                  {!contents.isLoading && (
+                    <span className="ml-2 text-xs opacity-70">
+                      {allItems.filter((item) => matchesOwnerRole(item.owner, tab.key)).length}
+                    </span>
+                  )}
+                </button>
+              ))}
+            </div>
+          )}
+
+          <div className="flex flex-col gap-3 sm:flex-row sm:items-end">
+            <div className="relative flex-1">
+              <Search className="absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2 text-muted-foreground" />
+              <input
+                type="text"
+                placeholder="Search by filename or URL..."
+                value={search}
+                onChange={(e) => setSearch(e.target.value)}
+                className="w-full rounded border border-border bg-background py-2 pl-10 pr-4 text-sm focus:outline-none focus:ring-2 focus:ring-hanover-green"
+              />
+            </div>
+            <select
+              value={statusFilter}
+              onChange={(e) => setStatusFilter(e.target.value)}
+              className="rounded border border-border bg-background px-4 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-hanover-green sm:min-w-40"
+            >
+              <option value="">All Statuses</option>
+              <option value="Created">Created</option>
+              <option value="in-progress">In Progress</option>
+              <option value="Finalized">Finalized</option>
+              <option value="Archived">Archived</option>
+            </select>
+            <select
+              value={typeFilter}
+              onChange={(e) => setTypeFilter(e.target.value)}
+              className="rounded border border-border bg-background px-4 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-hanover-green sm:min-w-40"
+            >
+              <option value="">All Types</option>
+              <option value="Reference">Reference</option>
+              <option value="Workflow">Workflow</option>
+            </select>
+            <Link
+              to="/hero/content/new"
+              className="flex shrink-0 items-center justify-center gap-2 rounded bg-hanover-green px-4 py-2 text-sm font-semibold text-white transition-colors hover:bg-hanover-green/90"
+            >
+              <Plus className="h-4 w-4" />
+              New Content
+            </Link>
           </div>
-          <select
-            value={status}
-            onChange={(e) => setStatus(e.target.value)}
-            className="rounded border border-border bg-background px-4 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-hanover-green sm:min-w-40"
-          >
-            <option value="">All Statuses</option>
-            <option value="Created">Created</option>
-            <option value="in-progress">In Progress</option>
-            <option value="Finalized">Finalized</option>
-            <option value="Archived">Archived</option>
-          </select>
-          <Link
-            to="/hero/content/new"
-            className="flex shrink-0 items-center justify-center gap-2 rounded bg-hanover-green px-4 py-2.5 text-sm font-semibold text-white transition-colors hover:bg-hanover-green/90"
-          >
-            <Plus className="h-4 w-4" />
-            New Content
-          </Link>
-        </div>
-        <div>
-          <select
-            className="mb-5 rounded border border-border bg-background px-4 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-hanover-green sm:min-w-40"
-            value={type}
-            onChange={(click) => setType(click.target.value)}
-          >
-            <option value="">No Filter</option>
-            <option value="Reference">Reference</option>
-            <option value="Workflow">Workflow</option>
-          </select>
         </div>
 
         {contents.isLoading ? (
@@ -90,27 +137,22 @@ function ContentListPage() {
             <p className="mt-2 break-words text-sm text-muted-foreground">
               {contents.error instanceof Error ? contents.error.message : String(contents.error)}
             </p>
-            <p className="mt-4 text-xs text-muted-foreground">
-              API should be at{" "}
-              {import.meta.env.VITE_API_URL ?? "http://127.0.0.1:3000 (dev default)"}. Run{" "}
-              <code className="rounded bg-muted px-1">pnpm dev</code> from the repo root.
-            </p>
           </div>
-        ) : contents.data?.length === 0 ? (
+        ) : filtered.length === 0 ? (
           <div className="py-16 text-center text-muted-foreground">No content found.</div>
         ) : (
           <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
-            {contents.data?.map((item) => (
+            {filtered.map((item) => (
               <Link
                 key={item.fileID}
                 to={`/hero/content/${item.fileID}/edit`}
                 className="group rounded border border-border bg-card p-5 shadow-sm transition-all hover:border-hanover-green hover:shadow-md"
               >
-                <div className="mb-3 flex items-start justify-between">
-                  <h3 className="text-base font-semibold text-foreground group-hover:text-hanover-green">
+                <div className="mb-3 flex items-start justify-between gap-2">
+                  <h3 className="text-base font-semibold leading-snug text-foreground group-hover:text-hanover-green">
                     {item.filename ?? "Untitled"}
                   </h3>
-                  <div className="ml-2 flex shrink-0 items-center gap-1">
+                  <div className="flex shrink-0 items-center gap-1">
                     <span
                       className={`rounded px-2 py-0.5 text-xs font-semibold ${getStatusBadge(item.document_status)}`}
                     >
@@ -167,4 +209,4 @@ function ContentListPage() {
   );
 }
 
-export default ContentListPage;
+export default ContentPage;
